lxd: Fix user lookup test for Fedora >= 34
A while ago we introduced a custom patch that was required because
the default nsswitch configuration of Fedora was depending on
sssd for local user lookup. As sssd is not running in the mock
build environment the error message was not as expected by the
upstream test. Fedora >= 34 reverted nsswitch back to files, the
custom patch doesn't need to be applied anymore.